requirements.txt
pyserial>=3.5
bleuio>=1.7.5
